refactor: separate out view switcher
