+++ desk/src/components/List/ListControls.vue
@@ -2,29 +2,28 @@
 	<div class="flex gap-2">
 		<ListFilters
 			v-if="options.showFilters"
-			v-model="config.filters"
-			:allFilterableFields="config.fields"
+			v-model="listConfig.filters"
+			:allFilterableFields="listConfig.fields"
 		/>
 		<ListColumnSettings
-			v-if="options.showColumnSettings && config.columns"
-			v-model="config.columns"
-			:allColumns="config.fields"
+			v-if="options.showColumnSettings && listConfig.columns"
+			v-model="listConfig.columns"
+			:allColumns="listConfig.fields"
 		/>
 		<ListSortSelector
 			v-if="options.showSortSelector"
-			v-model="config.sort"
-			:allSortableFields="config.columns"
+			v-model="listConfig.sort"
+			:allSortableFields="listConfig.columns"
 		/>
 	</div>
 </template>
 
 <script setup>
-import { defineModel } from "vue"
 import ListColumnSettings from "@/components/List/ListColumnSettings.vue"
 import ListFilters from "@/components/List/ListFilters.vue"
 import ListSortSelector from "@/components/List/ListSortSelector.vue"
 
-const config = defineModel()
+import { listConfig } from "@/stores/view"
 
 const emit = defineEmits(["update", "reload", "fetch"])
