+++ desk/src/stores/view.js
@@ -1,22 +1,40 @@
 import { createResource } from "frappe-ui"
-import { ref } from "vue"
+import { ref, computed } from "vue"
+import { isEqual } from "lodash"
 
-export const doctype = ref("")
-export const config_name = ref("")
-export const isDefaultConfig = ref(true)
-export const isDefaultOverriden = ref(false)
+const doctype = ref("")
+const configName = ref("")
+const isDefaultConfig = ref(true)
+const isDefaultOverriden = ref(false)
+const listConfig = ref({})
+const oldConfig = ref({})
+const configUpdated = computed(() => !isEqual(listConfig.value, oldConfig.value))
+const doctypeSavedViews = ref([])
 
-export const config_settings = createResource({
+const configSettings = createResource({
 	url: "frappe.desk.doctype.view_config.view_config.get_config",
 	makeParams() {
 		return {
 			doctype: doctype.value,
-			config_name: config_name.value,
+			config_name: configName.value,
 			is_default: isDefaultConfig.value,
 		}
 	},
-	onSuccess(response) {
-		isDefaultOverriden.value = !response.from_meta
-		config_name.value = response.name
+	onSuccess(data) {
+		isDefaultOverriden.value = !data.from_meta
+		configName.value = data.name
+		doctypeSavedViews.value = configSettings.data.views
 	},
 })
+
+export {
+	doctype,
+	configName,
+	configSettings,
+	listConfig,
+	oldConfig,
+	configUpdated,
+	isDefaultConfig,
+	isDefaultOverriden,
+	doctypeSavedViews,
+}
